requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
cohere==4.40
farm-haystack[docstores]==1.20.1